requirements.txt
# Core dependencies
click>=8.1.8
docker>=7.1.0
rich>=13.9.4
pyyaml>=6.0.2
requests>=2.32.3
pandas>=2.2.3
plotly>=6.0.1
jinja2>=3.1.6
pydantic>=2.11.0
typer>=0.15.2

# Security tools wrappers
pygrype>=0.4.0
trivyscan>=0.0.3b0
# syft>=0.1.0

# Testing
pytest>=8.3.5
pytest-cov>=6.0.0
black>=25.1.0
isort>=6.0.1
flake8>=4.0.0
mypy>=1.15.0

# Documentation
mkdocs>=1.6.1
mkdocs-material>=9.6.9